[vfs] Move some files out of BBGE, fixes to ttvfs code, add C-like wrapper API
ByteBuffer will be necessary for a few changes in other files.
The added C-like VFS API can be enabled with #define BBGE_BUILD_VFS.
If this is not defined, the normal fopen(), ... functions will be used
instead.
The next step is to change all fopen() and related calls, as well as
std::ifstream throughout the code, to use vfopen() or InStream.